# docs/openapi/generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@livetale.com
    name: LiveTale Team
  description: |
    LiveTale �꾨줈�앺듃�� API 紐낆꽭��
    �명꽣�숉떚釉� �ㅽ넗由ы뀛留곴낵 罹먮┃�� 梨꾪똿�� �꾪븳 醫낇빀 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LiveTale API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.livetale.com/v1
- description: Staging server
  url: https://staging-api.livetale.com/v1
- description: Development server
  url: http://localhost:3000/v1
security:
- BearerAuth: []
- ApiKeyAuth: []
paths:
...
  /notifications:
    get:
      description: �ъ슜�먯쓽 �뚮┝怨� 怨듭��ы빆 紐⑸줉�� 議고쉶�⑸땲��
      parameters:
      - description: �뚮┝ ���� �꾪꽣
        explode: true
        in: query
        name: type
        required: false
        schema:
          enum:
          - info
          - warning
          - achievement
          - story_update
          type: string
        style: form
      - description: �쎌� �딆� �뚮┝留� 議고쉶
        explode: true
        in: query
        name: unread_only
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: �섏씠吏� 踰덊샇
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: �섏씠吏��� ��ぉ ��
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_notifications_get_200_response"
          description: �뚮┝ 紐⑸줉 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �뚮┝/怨듭� 議고쉶
      tags:
      - Notification
      x-accepts:
      - application/json
      x-tags:
      - tag: Notification
  /notifications/{id}/read:
    patch:
      description: �뱀젙 �뚮┝�� �쎌쓬 泥섎━�⑸땲��
      parameters:
      - description: �뚮┝ ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_notifications__id__read_patch_200_response"
          description: �뚮┝ �쎌쓬 泥섎━ �깃났
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �섎せ�� �붿껌
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �뚮┝�� 李얠쓣 �� �놁쓬
      summary: �뚮┝ �쎌쓬 泥섎━
      tags:
      - Notification
      x-accepts:
      - application/json
      x-tags:
      - tag: Notification
components:
  schemas:
...
    Notification:
      example:
        is_read: true
        user_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        message: message
        type: info
      properties:
        id:
          format: uuid
          type: string
        user_id:
          format: uuid
          type: string
        title:
          type: string
        message:
          type: string
        type:
          enum:
          - info
          - warning
          - achievement
          - story_update
          type: string
        is_read:
          type: boolean
        created_at:
          format: date-time
          type: string
      type: object
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      type: object
...
    _worlds_get_200_response_pagination:
      example:
        total: 1
        limit: 6
        page: 0
        total_pages: 5
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      type: object
...
    _notifications_get_200_response:
      example:
        unread_count: 0
        pagination:
          total: 1
          limit: 6
          page: 0
          total_pages: 5
        notifications:
        - is_read: true
          user_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          message: message
          type: info
        - is_read: true
          user_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          message: message
          type: info
      properties:
        notifications:
          items:
            $ref: "#/components/schemas/Notification"
          type: array
        unread_count:
          description: �쎌� �딆� �뚮┝ ��
          type: integer
        pagination:
          $ref: "#/components/schemas/_worlds_get_200_response_pagination"
      type: object
    _notifications__id__read_patch_200_response:
      example:
        notification:
          is_read: true
          user_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          message: message
          type: info
        message: �뚮┝�� �쎌쓬 泥섎━�섏뿀�듬땲��
      properties:
        message:
          example: �뚮┝�� �쎌쓬 泥섎━�섏뿀�듬땲��
          type: string
        notification:
          $ref: "#/components/schemas/Notification"
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey